Rename app to http_app
