+++ Dockerfile
@@ -12,4 +12,4 @@ RUN poetry install --no-root
 
 COPY . .
 
-CMD ["uvicorn", "app:create_app", "--factory","--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "http_app:create_app", "--factory","--host", "0.0.0.0", "--port", "8000"]
