+++ README.md
@@ -12,7 +12,7 @@ This is an example implementation of a API service applying
 concepts from [Clean Architecture](https://blog.cleancoder.com/uncle-bob/2012/08/13/the-clean-architecture.html)
 and [SOLID principles](https://en.wikipedia.org/wiki/SOLID).
 
-* The book domain is isolated behind an interface class, enforcing the [Interface Segregation principle](https://en.wikipedia.org/wiki/Interface_segregation_principle) 
+* The books domain is isolated behind an interface class, enforcing the [Interface Segregation principle](https://en.wikipedia.org/wiki/Interface_segregation_principle) 
   and the [Inversion of Control principle](https://en.wikipedia.org/wiki/Inversion_of_control)
 * The same principles are used for the BookRepository class
 * The application framework is decoupled from the domain logic (we can potentially add a new application and reuse the same domain logic to provide e.g. a GRPC interface)
@@ -22,7 +22,7 @@ In this way our components are loosely coupled and the application logic
 (the domains package) is completely independent of from the chosen framework
 and the persistence layer.
 
-## API Docs
+## HTTP API Docs
 
 This application uses [fastapi-versionizer](https://github.com/alexschimpf/fastapi-versionizer)
 to provide easy API schema version management.
@@ -43,7 +43,7 @@ Packages are ordered from the highest level to the lowest one.
 
 ------
 
-* `app` (http presentation layer)
+* `http_app` (http presentation layer)
 * `storage` (database connection manager, repository implementation)
 
 ------
