--- app/templates.py
@@ -1,3 +1,0 @@
-from starlette.templating import Jinja2Templates
-
-templates = Jinja2Templates(directory="app/templates")
