+++ docker-compose.yaml
@@ -6,7 +6,7 @@ services:
       - '8000:8000'
     command:
       - 'uvicorn'
-      - 'app:create_app'
+      - 'http_app:create_app'
       - '--host'
       - '0.0.0.0'
       - '--port'
