+++ http_app/containers.py
@@ -2,7 +2,7 @@ from dependency_injector.containers import DeclarativeContainer, WiringConfigura
 from dependency_injector.providers import ThreadSafeSingleton, Dependency
 from sqlalchemy_bind_manager import SQLAlchemyBindManager
 
-from app import AppConfig
+from http_app import AppConfig
 from domains.books import BookServiceInterface
 from domains.books._local import LocalBookService
 from domains.books._local.data_access_interfaces import BookRepositoryInterface
@@ -16,8 +16,8 @@ class Container(DeclarativeContainer):
     Docs: https://python-dependency-injector.ets-labs.org/
     """
 
-    # Enable injection on the whole app package
-    wiring_config = WiringConfiguration(packages=["app", "storage", "domains"])
+    # Enable injection on the whole http_app package
+    wiring_config = WiringConfiguration(packages=["http_app", "storage", "domains"])
 
     """
     We could use the config provider but it would transform our nice typed
@@ -44,7 +44,7 @@ class Container(DeclarativeContainer):
     
     e.g.
     Mapping
-        MyInterface = providers.Factory("app.storage.repositories.ConcreteClass")
+        MyInterface = providers.Factory("http_app.storage.repositories.ConcreteClass")
 
     Usage
         @inject
