+++ http_app/routes/ping.py
