+++ http_app/templates.py
@@ -1,0 +1,3 @@
+from starlette.templating import Jinja2Templates
+
+templates = Jinja2Templates(directory="http_app/templates")
